feat: [OI-249] login api status
- adds react-query
- refactors api
- refactors useLoginData
- adds spid skeletons
- reduces code repeatition
  SpidSelection & SpidModal
- login refactoring
- imporeves error handling
