+++ src/oneid/oneid-ecs-core/src/main/webui/package.json
@@ -10,6 +10,7 @@
     "@mui/lab": "^5.0.0-alpha.170",
     "@mui/material": "^5.15.20",
     "@pagopa/mui-italia": "^1.4.2",
+    "@tanstack/react-query": "^5.62.2",
     "@types/node": "^20.14.5",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
