+++ src/oneid/oneid-ecs-core/src/main/webui/src/index.tsx
@@ -5,16 +5,20 @@ import { theme } from '@pagopa/mui-italia';
 
 import App from './App';
 import './locale';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 const rootEl = document.getElementById('root');
 
 if (rootEl) {
   const root = createRoot(rootEl);
+  const queryClient = new QueryClient();
 
   root.render(
     <ThemeProvider theme={theme}>
       <React.StrictMode>
-        <App />
+        <QueryClientProvider client={queryClient}>
+          <App />
+        </QueryClientProvider>
       </React.StrictMode>
     </ThemeProvider>
   );
