+++ src/oneid/oneid-ecs-core/src/main/webui/src/pages/login/components/SpidButton/SpidButton.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { vi } from 'vitest';
-import { SpidButton, SpidButtonProps } from './SpidButton';
-import { i18nTestSetup } from '../../../__tests__/i18nTestSetup';
+import { SpidButton, SpidButtonProps } from '../SpidButton';
+import { i18nTestSetup } from '../../../../__tests__/i18nTestSetup';
 
 const TEST_ID = 'spidButton';
 
@@ -22,23 +22,15 @@ describe('SpidButton Component', () => {
     render(<SpidButton {...props} />);
 
   it('renders the button with the correct text', () => {
-    renderComponent({ loading: false, onClick: onClickMock });
+    renderComponent({ onClick: onClickMock });
 
     const button = screen.getByTestId(TEST_ID);
     expect(button).toBeInTheDocument();
     expect(button).toHaveTextContent('Login with SPID');
   });
 
-  it('renders the button in a loading state', () => {
-    renderComponent({ loading: true, onClick: onClickMock });
-
-    const button = screen.getByTestId(TEST_ID);
-    expect(button).toHaveAttribute('aria-busy', 'true');
-    expect(button).toBeDisabled();
-  });
-
   it('triggers the onClick function when clicked', () => {
-    renderComponent({ loading: false, onClick: onClickMock });
+    renderComponent({ onClick: onClickMock });
 
     const button = screen.getByTestId(TEST_ID);
     fireEvent.click(button);
@@ -47,7 +39,7 @@ describe('SpidButton Component', () => {
   });
 
   it('displays the SPID icon', () => {
-    renderComponent({ loading: false, onClick: onClickMock });
+    renderComponent({ onClick: onClickMock });
 
     const icon = screen.getByAltText('SPID Icon');
     expect(icon).toBeInTheDocument();
