+++ src/oneid/oneid-ecs-core/src/main/webui/src/pages/login/components/SpidModal/SpidModal.test.tsx
@@ -1,12 +1,12 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 
-import { IdentityProviders } from '../../../utils/IDPS';
-import { trackEvent } from '../../../services/analyticsService';
-import { forwardSearchParams } from '../../../utils/utils';
-import SpidModal from './SpidModal';
+import { IdentityProviders } from '../../../../utils/IDPS';
+import { trackEvent } from '../../../../services/analyticsService';
+import { forwardSearchParams } from '../../../../utils/utils';
+import SpidModal from '.././SpidModal';
 
-vi.mock('../../../services/analyticsService');
-vi.mock('../../../utils/utils', () => ({
+vi.mock('../../../../services/analyticsService');
+vi.mock('../../../../utils/utils', () => ({
   forwardSearchParams: vi.fn(() => 'testParams'),
 }));
 
@@ -25,6 +25,7 @@ describe('SpidModal', () => {
   it('renders the modal with identity providers', () => {
     render(
       <SpidModal
+        loading={false}
         openSpidModal={true}
         setOpenSpidModal={vi.fn()}
         idpList={mockIdpList}
