+++ src/oneid/oneid-ecs-core/src/main/webui/src/pages/login/components/SpidSelect/SpidSelect.test.tsx
@@ -2,8 +2,8 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { vi } from 'vitest';
 
-import { ENV } from '../../../utils/env';
-import SpidSelect from './SpidSelect';
+import { ENV } from '../../../../utils/env';
+import SpidSelect from '../SpidSelect';
 
 const oldWindowLocation = global.window.location;
 beforeAll(() => {
